Awkward merge - probably some whitespace issues.  Adds some intelligence to SceneNode and prepares to incorporate Renderables.

+++ Engine/OrthoCamera.cpp
@@ -3,5 +3,5 @@
 OrthoCamera::OrthoCamera(const std::string &name): Camera(name), _zoom(0.5), _near(-10.0), _far(10.0) {}
 
 void OrthoCamera::setAspectRatio(const float ratio) {
-	setProjection(Matrix4::MakeOrtho(-(ratio/_zoom), (ratio/_zoom), -1.0 / _zoom, 1.0 / _zoom, _near, _far));
+	setProjection(Matrix4::MakeOrtho(-(ratio/_zoom), (ratio/_zoom), -(float)1.0 / _zoom, (float)1.0 / _zoom, _near, _far));
 }
